+++ devtools/src/lib/config.ts
@@ -68,8 +68,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.0.6",
-  genVersion: "2.605.6",
+  sdkVersion: "0.0.7",
+  genVersion: "2.607.1",
   userAgent:
-    "speakeasy-sdk/typescript 0.0.6 2.605.6 1.0.0 @avalanche-sdk/devtools",
+    "speakeasy-sdk/typescript 0.0.7 2.607.1 1.0.0 @avalanche-sdk/devtools",
 } as const;